Move sendReferralEmail to routes instead of mutation
