+++ lib/server/graphql/index.js
@@ -5,9 +5,6 @@
  */
 
 const graphql = require('graphql');
-const StorjMailer = require('storj-service-mailer');
-const defaults = require('../../config.js').DEFAULTS;
-const mailer = new StorjMailer(defaults.mailer);
 const Promise = require('bluebird');
 
 let _schema;
@@ -63,20 +60,6 @@ module.exports = {
         console.error(`Unhandled exception in \`addPaymentProcessor\`: ${e}`);
         console.error(e.stack);
       });
-  },
-  sendReferralEmail: function(senderEmail, recipientEmail, marketing) {
-    return new Promise((resolve, reject) => {
-      mailer.dispatch(recipientEmail, 'referral', {
-        url: 'https://app.storj.io/#/signup?referralLink=' + marketing.referralLink,
-        senderEmail: senderEmail
-      }, function(err) {
-        if (err) {
-          console.error('Error sendReferralEmail: ', err);
-          return reject(err);
-        }
-        return resolve(true);
-      });
-    });
   },
   get schema() {
     if (!_schema) {
