+++ lib/server/graphql/mutations/index.js
@@ -3,14 +3,12 @@
 const graphql = require('graphql');
 const addPaymentMethod = require('./add-payment-method');
 const removePaymentMethod = require('./remove-payment-method');
-const sendReferralEmails = require('./send-referral-emails');
 
 const rootMutation = new graphql.GraphQLObjectType({
   name: 'Mutation',
   fields: {
     addPaymentMethod,
-    removePaymentMethod,
-    sendReferralEmails
+    removePaymentMethod
   }
 });
