+++ lib/server/routefactory.js
@@ -5,7 +5,7 @@ module.exports = function RouteFactory(options) {
     require('./routes/credits'),
     require('./routes/debits'),
     require('./routes/users'),
-    require('./routes/graphql')
+    require('./routes/graphql'),
     require('./routes/referrals')
   ]).map(function(Router) {
     return Router({
