+++ lib/server/routes/credits.js
@@ -319,9 +319,9 @@ CreditsRouter.prototype.confirmCredit = function(req, res) {
 };
 
 CreditsRouter.prototype.getReferralSenderCreditsByDate = function(req, res, next) {
-  const Credit = this.models.credit; 
-  Credit.find({ 
-    created: { $gte: req.periodStart, $lte: req.periodEnd } 
+  const Credit = this.models.credit;
+  Credit.find({
+    created: { $gte: req.periodStart, $lte: req.periodEnd }
   }, function(err, credits) {
     if (err) {
       return next(new errors.InternalError(err.message));
@@ -348,8 +348,8 @@ CreditsRouter.prototype._definitions = function() {
       this.checkType(/^invoice.payment_(succeeded|failed)$/),
       this.confirmCredit
     ],
-    ['GET', 
-     '/credits/referral/:periodStart/:periodEnd', 
+    ['GET',
+     '/credits/referral/:periodStart/:periodEnd',
      authenticate,
      this.getReferralSenderCreditsByDate
     ],
